// frontend/src/lib/api/generated/models/TicketTypeCode.ts
/**
 * Roominar
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 0.1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { HttpFile } from '../http/http';

export enum TicketTypeCode {
    EarlyBird = 'EARLY_BIRD',
    Vip = 'VIP',
    Group = 'GROUP',
    Corporate = 'CORPORATE',
    Student = 'STUDENT',
    Free = 'FREE',
    Donation = 'DONATION',
    Multiday = 'MULTIDAY',
    DayPass = 'DAY_PASS',
    Regular = 'REGULAR',
    LateBird = 'LATE_BIRD',
    Online = 'ONLINE',
    Onsite = 'ONSITE',
    Press = 'PRESS',
    Speaker = 'SPEAKER',
    Staff = 'STAFF',
    Exhibitor = 'EXHIBITOR',
    Sponsor = 'SPONSOR',
    AccessPass = 'ACCESS_PASS',
    SeasonPass = 'SEASON_PASS',
    Discounted = 'DISCOUNTED',
    Alumni = 'ALUMNI',
    Family = 'FAMILY',
    Child = 'CHILD',
    EarlyAccess = 'EARLY_ACCESS',
    WorkshopPass = 'WORKSHOP_PASS',
    Official = 'OFFICIAL'
}